fix: borderRadius and clipper can't be both null
In `ClipRRect` constructor, `borderRadius` and `clipper` are asserted
that they shouldn't be null simultaneously. When `clipper` is not used
`borderRadius` can't has null value. The solution is to provide it a
default one.

If this issue not fixed, calling without `borderRadius` parameter will
encounter failed assertion. `assert(borderRadius != null || clipper !=
null)`

References

* https://api.flutter.dev/flutter/widgets/ClipRRect/ClipRRect.html